Update Authy - Download - Setup - cURL.bat

+++ Authy - Download - Setup - cURL.bat
@@ -1,7 +1,7 @@
 @ECHO OFF
 title _Hieuck.IT_'s Windows Application
 color 0B
-mode con:cols=100 lines=20
+mode con:cols=100 lines=17
 @cls
 echo.
 echo.
@@ -15,42 +15,92 @@ echo.
 @echo                 Dang Cai Dat Authy. Vui Long Cho
 @echo off
 pushd "%~dp0"
-
 :: Set File Name Link User Agent
 set "FILENAME=Authy-HieuckIT.exe"
 set "LINK64=https://electron.authy.com/download?channel=stable&arch=x64&platform=win32&version=latest&product=authy"
 set "LINK32=https://electron.authy.com/download?channel=stable&arch=x32&platform=win32&version=latest&product=authy"
 set "USERAGENT=Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/58.0.3029.110 Safari/537.3"
 
+:: Check if Command Prompt is running with administrator privileges
+net session >nul 2>&1
+if %errorlevel% == 0 (
+	echo Command Prompt is running as Administrator.
+) else (
+	echo Command Prompt is not running as Administrator.
+	echo Please Run as Administrator. Exiting in 5 seconds...
+	for /l %%i in (5,-1,1) do (
+		echo Exiting in %%i seconds...
+		timeout /t 1 /nobreak >nul
+	)
+	exit
+)
+
 :: Terminate the Authy Process
 tasklist | find /i "Authy Desktop.exe" > nul
 if %errorlevel% equ 0 (
-    taskkill /im "Authy Desktop.exe" /f
+	taskkill /im "Authy Desktop.exe" /f
 )
 
 :: Detect Windows Architecture
 if exist "%SYSTEMROOT%\SysWOW64" (
-    set "ARCH=x64"
+	set "ARCH=x64"
 ) else (
-    set "ARCH=x86"
+	set "ARCH=x86"
 )
 
 :: Download
+@ECHO OFF
+title _Hieuck.IT_'s Windows Application
+color 0B
+mode con:cols=100 lines=17
+@cls
+echo.
+echo.
+echo.
+@echo     Бл          ллл   ллл ллл лллллл ллл   ллл  ллллл  ллл  лл ллл ллллллллл
+@echo       л         ллл   ллл ллл ллл    ллл   ллл ллл   л ллл лл  ллл    ллл
+@echo        Вл       ллллллллл ллл лллллл ллл   ллл ллл     ллллл   ллл    ллл
+@echo       л         ллл   ллл ллл ллл    ллл   ллл ллл   л ллл лл  ллл    ллл
+@echo     Бл   ВВВВВ  ллл   ллл ллл лллллл ллллллллл  ллллл  ллл  лл ллл    ллл В
+@echo.  
+@echo                 Dang Cai Dat Authy. Vui Long Cho
+@echo off
+pushd "%~dp0"
 echo Downloading Authy...
 if %ARCH%==x64 (
-    curl --insecure -L --max-redirs 20 -A "%USERAGENT%" -o "%FILENAME%" "%LINK64%"
+	curl --insecure -L --max-redirs 20 -A "%USERAGENT%" -o "%FILENAME%" "%LINK64%"
 ) else (
-    curl --insecure -L --max-redirs 20 -A "%USERAGENT%" -o "%FILENAME%" "%LINK32%"
+	curl --insecure -L --max-redirs 20 -A "%USERAGENT%" -o "%FILENAME%" "%LINK32%"
 )
 
 if not exist "%FILENAME%" (
-    echo Download Authy failed.
-    echo Please check your network connection. Exiting in 5 seconds...
-    timeout /t 5 /nobreak >nul
-    exit
+	echo Download Authy failed.
+	echo Please check your network connection. Exiting in 5 seconds...
+	for /l %%i in (5,-1,1) do (
+		echo Exiting in %%i seconds...
+		timeout /t 1 /nobreak >nul
+	)
+	exit
 )
 
 :: Install
+@ECHO OFF
+title _Hieuck.IT_'s Windows Application
+color 0B
+mode con:cols=100 lines=17
+@cls
+echo.
+echo.
+echo.
+@echo     Бл          ллл   ллл ллл лллллл ллл   ллл  ллллл  ллл  лл ллл ллллллллл
+@echo       л         ллл   ллл ллл ллл    ллл   ллл ллл   л ллл лл  ллл    ллл
+@echo        Вл       ллллллллл ллл лллллл ллл   ллл ллл     ллллл   ллл    ллл
+@echo       л         ллл   ллл ллл ллл    ллл   ллл ллл   л ллл лл  ллл    ллл
+@echo     Бл   ВВВВВ  ллл   ллл ллл лллллл ллллллллл  ллллл  ллл  лл ллл    ллл В
+@echo.  
+@echo                 Dang Cai Dat Authy. Vui Long Cho
+@echo off
+pushd "%~dp0"
 echo Installing Authy...
 "%FILENAME%" /S
 
@@ -58,8 +108,11 @@ echo Installing Authy...
 if not exist "%LocalAppData%\Authy\Authy Desktop.exe" (
 	echo Installation Authy failed.
 	echo Please try Run as Administrator. Exiting in 5 seconds...
-	timeout /t 5 /nobreak >nul
-    exit
+	for /l %%i in (5,-1,1) do (
+		echo Exiting in %%i seconds...
+		timeout /t 1 /nobreak >nul
+	)
+	exit
 ) else (
 	echo Installation Authy complete.
 )
@@ -67,5 +120,8 @@ if not exist "%LocalAppData%\Authy\Authy Desktop.exe" (
 :: Clean Up
 del "%FILENAME%"
 echo The script will automatically close in 5 seconds.
-timeout /t 5 /nobreak >nul
+for /l %%i in (5,-1,1) do (
+	echo Closing in %%i seconds...
+	timeout /t 1 /nobreak >nul
+)
 popd
